chore: Rename create-modal to modal to better represent its purpose
