+++ frontend/src/app/features/calendar/components/calendar-main/calendar-main.component.html
@@ -56,7 +56,7 @@
 
   <div class="flex-1">
     <!--  Modal-->
-    <app-create-modal>
+    <app-modal>
       <div modal-header>
         <div class="join">
           <button
@@ -81,7 +81,7 @@
       } @else {
         <app-event-form />
       }
-    </app-create-modal>
+    </app-modal>
 
     <!-- Router Outlet for Calendar Views -->
     <router-outlet></router-outlet>
