+++ frontend/src/app/shared/components/modal/service/modal.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from "@angular/core/testing";
 
-import { CreateModalService } from "./create-modal.service";
+import { ModalService } from "./modal.service";
 
 describe("CreateModalService", () => {
-  let service: CreateModalService;
+  let service: ModalService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(CreateModalService);
+    service = TestBed.inject(ModalService);
   });
 
   it("should be created", () => {
